News feed video resizes correctly on different resolutions
Main change was the news feed livestream now changes size depending on the size of the panel it is in. Currently it uses preset sizes defined, and has 3 different 16:9 ratios (I can add more if needed when testing on monitors if need be). Reason I don't want to do this size setting dynamically is because the video frame needs to stay 16:9 ratio and this is the best way I could do it for the time being.

Small other changes:
- Icon in tab is now the IUP hawk that is used in top left corner of toolbar header, just resized to be a lot smaller.
- Added jQuery plugin to easily detect when screen is in full-screen mode, in case it needs to be used for other things besides the news live feed (which I'm going to try and use it for the Twitter and Facebook plugins to make them dynamically more reliable when resizing).
- More CSS changes to allow the news feed iframe to be adjusted accordingly.